Search entire window title in Linux findWindow()
The LinuxUtil.java 'findWindow()' uses "wmctrl -lpGx" to get info
about the window.  It splits the result on any whitespace, and then
trys to match the given appName against fields 7 or 9.  However, if
the title has spaces then only the first word is in field 9.  Changing
the split to have a limit results in having the entire title
available.  Also, it makes more sense to do the 'contains' search on
field 9 (the title) rather than field 7 (the "class").